+++ conversation/src/main/java/me/khw7385/conversation/infrastructure/websocket/client/dto/ClientToServerAudioMessage.java
@@ -1,4 +1,4 @@
-package me.khw7385.conversation.infrastructure.websocket.client;
+package me.khw7385.conversation.infrastructure.websocket.client.dto;
 
 public record ClientToServerAudioMessage(
         String audio
