+++ conversation/src/main/java/me/khw7385/conversation/infrastructure/websocket/client/dto/ServerToClientAudioMessage.java
@@ -1,4 +1,4 @@
-package me.khw7385.conversation.infrastructure.websocket.client;
+package me.khw7385.conversation.infrastructure.websocket.client.dto;
 
 import me.khw7385.conversation.application.port.outbound.Message;
